fix: start showing events from the current date
The previous implementation had an undesired behaviour, it showed events from the first month with events no matter what the current month was.

In this new implementation, we check if the current month is in the list and use it as the starting point. If not, we use the first month with events.

In the future we'll want a more sophisticated way of doing this, for example, rendering the monthly calendar regardless of the events list for that month.